[DEV-154] rename: todo 컴포넌트 components 디렉토리로 이동
