+++ src/components/goal-detail/todo-list.tsx
@@ -9,7 +9,7 @@ import { useInfiniteScroll } from "@/hooks/use-infinite-scroll";
 import { cn } from "@/lib/cn";
 import { getInfiniteTodosByGoalIdOptions } from "@/services/todo/query";
 
-import Todo from "../@common/todo";
+import Todo from "../todo/todo";
 import TodoModal from "../todo-modal/todo-modal";
 
 interface TodoListProps {
