+++ src/components/todo/all-list-todo.tsx
@@ -5,7 +5,7 @@ import { useState } from "react";
 import { useSuspenseInfiniteQuery } from "@tanstack/react-query";
 
 import PlusIcon from "@/assets/icons-small/plus/plus_db_sm.svg";
-import Todo from "@/components/@common/todo";
+import Todo from "@/components/todo/todo";
 import { useInfiniteScroll } from "@/hooks/use-infinite-scroll";
 import { getInfiniteTodosByGoalIdOptions } from "@/services/todo/query";
